Refatora teste de view de detalhe de dataset para não ficar acoplada à app covid19
